+++ src/main/java/com/example/emptySaver/repository/ScheduleRepository.java
@@ -9,4 +9,5 @@ import java.util.List;
 public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
 
     List<Schedule> findByPublicType(boolean publicType);
+    List<Schedule> findByOriginScheduleId(Long originScheduleId);
 }
